Remove unused icon files and update icon references in the application
- Deleted the old icon files (icon.ico and electron.svg) to streamline asset management.
- Replaced references to the removed electron.svg with the new icon.svg in AppLayout component.
- Updated the download count parsing in gameService to use parseFloat for improved accuracy.
- Added a new 'Popularity' column in GamesView to display download counts, enhancing user insights.

+++ src/main/services/gameService.ts
@@ -523,7 +523,7 @@ class GameService {
           size,
           lastUpdated,
           releaseName,
-          downloads: parseInt(downloads) || 0,
+          downloads: parseFloat(downloads) || 0,
           thumbnailPath: thumbnailExists ? thumbnailPath : '',
           isInstalled: false
         }
